MDL-59718 tool_dataprivacy: Enable the user to make data requests
* Add functionality that enables the user to make data requests to the
site's Data Protection Officer (DPO).
When the user lodges a data request, this fires an adhoc task that
prepares the user's metadata. Upon process completion, an email will be
sent to the user(s) matching the role of the Data Protection Officer.
Both an email and a Moodle notification will be sent to the DPO(s).